* Dummy model now return images with np.uint8 values.

+++ model.py
@@ -5,4 +5,4 @@ class Model:
     """Dummy model"""
     def generate(self, inputs):
         shape = (32, 128, 128, 3)
-        return np.random.uniform(size=np.prod(shape)).reshape(shape)
+        return np.random.uniform(size=np.prod(shape)).reshape(shape).astype(np.uint8)
